Update guide and playground
