+++ playground/vite/src/App.tsx
@@ -1,12 +1,33 @@
-import { Route, Routes } from 'react-router-dom';
-import DashboardPage from './pages';
-import OrdersPage from './pages/orders';
+import DashboardIcon from '@mui/icons-material/Dashboard';
+import ShoppingCartIcon from '@mui/icons-material/ShoppingCart';
+import { Outlet } from 'react-router-dom';
+import { AppProvider } from '@toolpad/core/react-router-dom';
+import type { Navigation } from '@toolpad/core';
+
+const NAVIGATION: Navigation = [
+  {
+    kind: 'header',
+    title: 'Main items',
+  },
+  {
+    title: 'Dashboard',
+    icon: <DashboardIcon />,
+  },
+  {
+    segment: 'orders',
+    title: 'Orders',
+    icon: <ShoppingCartIcon />,
+  },
+];
+
+const BRANDING = {
+  title: 'My Toolpad Core App',
+};
 
 export default function App() {
   return (
-    <Routes>
-      <Route path="/" Component={DashboardPage} />
-      <Route path="/orders" Component={OrdersPage} />
-    </Routes>
+    <AppProvider navigation={NAVIGATION} branding={BRANDING}>
+      <Outlet />
+    </AppProvider>
   );
 }
